Final Bone weights Linear Normalization

+++ src/COLLADA2GLTFWriter.cpp
@@ -1625,24 +1625,24 @@ bool COLLADA2GLTF::Writer::writeController(const COLLADAFW::Controller* controll
 			maxweight = *weights[0];
 			minweight = *weights[0];
 		}
-		for(int i = 0; i < weights.size(); i++) {
+		for(size_t i = 0; i < weights.size(); i++) {
 			maxweight = std::max(maxweight, *weights[i]);
 			minweight = std::min(minweight, *weights[i]);
 		}
 		if (maxweight == minweight) {
 			if (maxweight == 0) {
-				for(int i = 0; i < weights.size(); i++) {
+				for(size_t i = 0; i < weights.size(); i++) {
 					*weights[i] = 0;
 				}
 			}
 			else {
-				for(int i = 0; i < weights.size(); i++) {
+				for(size_t i = 0; i < weights.size(); i++) {
 					*weights[i] = 1;
 				}
 			}
 		}
 		else {
-			for(int i = 0; i < weights.size(); i++) {
+			for(size_t i = 0; i < weights.size(); i++) {
 				*weights[i] = (*weights[i] - minweight)/(maxweight - minweight);
 			}
 		}
